requirements.txt
asgiref==3.11.0
django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.11
pyjwt==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
channels==4.2.0
daphne==4.1.2

